Crud for Category and Product (cont)

+++ src/main/java/com/example/demo/service/RoleServiceImpl.java
@@ -3,19 +3,24 @@ package com.example.demo.service;
 import com.example.demo.entity.Permission;
 import com.example.demo.entity.Role;
 import com.example.demo.entity.Staff;
+import com.example.demo.exception.RoleNotCreateByCurrentStaffException;
 import com.example.demo.exception.RoleNotFoundException;
 import com.example.demo.form.RoleForm;
 import com.example.demo.repository.RoleRepository;
 import com.example.demo.response.RoleResponse;
 import com.example.demo.response.SimpleRoleResponse;
 import com.example.demo.security.SecurityUtil;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import javax.persistence.Id;
 import java.util.List;
+import java.util.Optional;
 import java.util.Set;
 import java.util.stream.Collectors;
 
+@Slf4j
 @Service(value = "roleService")
 public class RoleServiceImpl implements RoleService {
 
@@ -28,7 +33,7 @@ public class RoleServiceImpl implements RoleService {
     public RoleServiceImpl(RoleRepository roleRepository,
                            SecurityUtil securityUtil,
                            StaffServiceImpl staffService,
-                           PermissionServiceImpl permissionService ) {
+                           PermissionServiceImpl permissionService) {
         this.roleRepository = roleRepository;
         this.securityUtil = securityUtil;
         this.staffService = staffService;
@@ -52,10 +57,10 @@ public class RoleServiceImpl implements RoleService {
 
     @Override
     public Role save(RoleForm roleForm) {
-        Staff createByStaff = staffService.findStaffByUsername(
+        Staff createByStaff = staffService.findByUsername(
                 securityUtil.getCurrentPrincipal().getUsername());
         Set<Permission> permissions = permissionService.findAllByIdIsIn(roleForm.getPermissions());
-        Role role = roleRepository.save(RoleForm.buildRole(roleForm, createByStaff, permissions));
+        Role role = roleRepository.save(RoleForm.buildRole(roleForm.getName(), createByStaff, permissions));
         return role;
     }
 
@@ -63,15 +68,27 @@ public class RoleServiceImpl implements RoleService {
     public Role update(Integer id, RoleForm roleForm) {
         Role role = roleRepository.findById(id)
                 .orElseThrow(() -> new RoleNotFoundException(id));
-        Staff createByStaff = staffService.findStaffByUsername(
-                securityUtil.getCurrentPrincipal().getUsername());
+        Staff createByStaff = securityUtil.getCurrentStaff();
         Set<Permission> permissions = permissionService.findAllByIdIsIn(roleForm.getPermissions());
         return roleRepository.save(Role.updateData(role, roleForm, createByStaff, permissions));
     }
 
     @Override
-    public String delete(Integer id) {
-        roleRepository.deleteById(id);
-        return String.valueOf(id);
+    public void delete(Integer id) {
+//        Admin can delete any role
+        Staff staff = securityUtil.getCurrentStaff();
+        if (staff.isRootAdmin()) {
+            roleRepository.deleteById(id);
+            return;
+        }
+
+//        Staff can only delete role that they created
+        Optional<Role> optionalRole = roleRepository.findByIdAndCreatedBy(id, staff);
+        if (optionalRole.isPresent()) {
+            roleRepository.deleteById(id);
+        } else {
+            throw new RoleNotCreateByCurrentStaffException();
+        }
     }
+
 }
